Disable code that does matrix multiply in bins.
Previously SvmLinearSolver::matMult attempted to partition rows into
bins of 8 * 1024, do the multiplication for each bin, and then sum the
bins.  However an error in operator precidence meant that the code
never actually ran.

Make the minimal number of changes to completely disable the logic.  This
leaves the code a mess, but tested.